Task 2. Named entity recognition + image classification/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
Pillow>=8.2.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
pandas>=1.3.0
numpy>=1.20.3
ipywidgets>=7.6.5
notebook>=6.4.0
